changed datatype
Converted the datatype of Course. prereq to list instead of set, because a dictionary cannot be directly added into a set and will result in TypeError: unhashable type: 'dict'. This can be solved by converting the dictionary into a frozenset or a tuple, however, it needs a lot of computation between type conversions.